requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
SQLAlchemy>=2.0.25
